Time: 0 ms (100%), Space: 8.3 MB (99.94%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int reverse(int x) {
-    long long rev=0;
+    double rev=0;
     int hold;
